Move KeyBinder to Common to allow its use in Editor/SP. Also allow GUI usage in Common (excluded in Server).
