+++ Alliance.Client/Extensions/AdminMenu/Views/AdminSystem.cs
@@ -1,5 +1,6 @@
-﻿using Alliance.Client.Core.KeyBinder.Models;
-using Alliance.Client.Extensions.AdminMenu.ViewModels;
+﻿using Alliance.Client.Extensions.AdminMenu.ViewModels;
+using Alliance.Common.Core.KeyBinder;
+using Alliance.Common.Core.KeyBinder.Models;
 using Alliance.Common.Core.Security.Extension;
 using Alliance.Common.Extensions.AdminMenu.NetworkMessages.FromClient;
 using System.Collections.Generic;
