+++ Alliance.Client/Patch/DirtyClientPatcher.cs
@@ -15,7 +15,6 @@ namespace Alliance.Client.Patch
 
 			//TODO : Following 1.2 -> Check if any patch can be removed
 			//patchSuccess &= Patch_MissionMultiplayerGameModeFlagDominationClient.Patch();
-			patchSuccess &= Patch_KeyBinder.Patch();
 			patchSuccess &= Patch_MissionNetworkComponent.Patch();
 			patchSuccess &= Patch_DefaultAdminPanelOptionProvider.Patch();
 			patchSuccess &= Patch_AllianceAgentVisualSpawnComponent.Patch();
