+++ Alliance.Editor/SubModule.cs
@@ -1,4 +1,5 @@
-﻿using Alliance.Common.Core.Utils;
+﻿using Alliance.Common.Core.KeyBinder;
+using Alliance.Common.Core.Utils;
 using Alliance.Common.Extensions.AdvancedCombat.Behaviors;
 using Alliance.Common.Extensions.AnimationPlayer;
 using Alliance.Common.GameModels;
@@ -13,6 +14,7 @@ using Alliance.Common.Utilities;
 using Alliance.Editor.GameModes.Story.Utilities;
 using Alliance.Editor.GameModes.Story.Views;
 using Alliance.Editor.Patch;
+using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 using TaleWorlds.Core;
@@ -33,6 +35,14 @@ namespace Alliance.Editor
 		{
 			Common.SubModule.CurrentModuleName = ModuleId;
 
+			// Register and initialize Key Binder
+			List<Assembly> assemblies = new List<Assembly>
+			{
+				Assembly.GetAssembly(typeof(Common.SubModule)),
+				Assembly.GetAssembly(typeof(Editor.SubModule))
+			};
+			KeyBinder.Initialize(assemblies);
+
 			ActionFactory.Initialize();
 			ScenarioManager.Instance = new ScenarioManager();
 			SceneList.Initialize();
@@ -47,7 +57,6 @@ namespace Alliance.Editor
 
 			// Apply Harmony patches
 			DirtyCommonPatcher.Patch();
-
 			DirtyEditorPatcher.Patch();
 
 			Log("Alliance.Editor initialized", LogLevel.Debug);
