+++ Alliance.sln
@@ -8,6 +8,11 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Alliance.Client", "Alliance.Client\Alliance.Client.csproj", "{4A36A725-0920-49A0-AE3E-5037A61EA05B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Alliance.Server", "Alliance.Server\Alliance.Server.csproj", "{D7819BB5-5122-4087-A8E2-F9546080C1BC}"
+	ProjectSection(ProjectDependencies) = postProject
+		{4A36A725-0920-49A0-AE3E-5037A61EA05B} = {4A36A725-0920-49A0-AE3E-5037A61EA05B}
+		{65C67AD9-1CD6-4336-95C3-6D664D100A4F} = {65C67AD9-1CD6-4336-95C3-6D664D100A4F}
+		{C211171F-9278-4C9E-994F-E9E5612A5D4E} = {C211171F-9278-4C9E-994F-E9E5612A5D4E}
+	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Alliance.Editor", "Alliance.Editor\Alliance.Editor.csproj", "{C211171F-9278-4C9E-994F-E9E5612A5D4E}"
 EndProject
